Restructured approach, added incomplete matching.
Created a new approach that first generates a list of all files found
in the reference folders, then searches through that. This replaces
naive approach used in the first iteration that simply tried to open
the search term input by the user.

Created the Reference class in reference.py that stores information
about the reference documents found in the user's search paths. This
greatly facilitates passing the information around, and abstracts away
some of the low level implementation.

Matching is now based on whether the filename of the reference contains
the search string input by the user. For now, if there are multiple hits
we just print out what the hits were, so that the user can refine their
search.

Finally, some basic doctests were added. Unit testing will be added in
the next commit as I learn how to use the unittest module.